requirements.txt
scikit-learn
pandas
azureml-mlflow
gast
glob2
imutils
numpy
joblib
argparse